Move remaining Pybullet code to dedicated modules (#1171)
- Using IDE functionality
