+++ src/pybullet_helpers/motion_planning.py
@@ -1,37 +1,15 @@
-"""Interfaces to PyBullet robots."""
+"""Motion Planning in Pybullet."""
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Collection, Iterator, List, Optional, \
-    Sequence
+from typing import Collection, Iterator, Optional, Sequence
 
 import numpy as np
 import pybullet as p
 
 from predicators.src import utils
+from predicators.src.pybullet_helpers.robots import SingleArmPyBulletRobot
 from predicators.src.settings import CFG
 from predicators.src.structs import JointsState
-
-if TYPE_CHECKING:
-    from predicators.src.pybullet_helpers.robots import SingleArmPyBulletRobot
-
-########################### Other utility functions ###########################
-
-
-def get_kinematic_chain(robot: int, end_effector: int,
-                        physics_client_id: int) -> List[int]:
-    """Get all of the free joints from robot base to end effector.
-
-    Includes the end effector.
-    """
-    kinematic_chain = []
-    while end_effector > -1:
-        joint_info = p.getJointInfo(robot,
-                                    end_effector,
-                                    physicsClientId=physics_client_id)
-        if joint_info[3] > -1:
-            kinematic_chain.append(end_effector)
-        end_effector = joint_info[-1]
-    return kinematic_chain
 
 
 def run_motion_planning(
