+++ src/pybullet_helpers/utils.py
@@ -1,0 +1,23 @@
+"""Other Pybullet utility functions."""
+from __future__ import annotations
+
+from typing import List
+
+import pybullet as p
+
+
+def get_kinematic_chain(robot: int, end_effector: int,
+                        physics_client_id: int) -> List[int]:
+    """Get all of the free joints from robot base to end effector.
+
+    Includes the end effector.
+    """
+    kinematic_chain = []
+    while end_effector > -1:
+        joint_info = p.getJointInfo(robot,
+                                    end_effector,
+                                    physicsClientId=physics_client_id)
+        if joint_info[3] > -1:
+            kinematic_chain.append(end_effector)
+        end_effector = joint_info[-1]
+    return kinematic_chain
